refact: Atualização de Bibliotecas e Package Corrigindo Package com nome errado Atualizando versão das bibliotecas e gradle.
